machine_learning/requirements.txt
Flask
flask_cors
numpy
pymongo
python-dotenv
tensorflow-cpu
gunicorn
Pillow
werkzeug
requests
psutil
fastapi
uvicorn[standard]
python-multipart
